weight/app/requirements.txt
flask
flask-restful 
mysql-connector-python